Clean up sockets
