+++ src/app/item-popup/ItemSocketsGeneral.tsx
@@ -1,4 +1,3 @@
-import { LockedItemType } from 'app/loadout-builder/types';
 import {
   CHALICE_OF_OPULENCE,
   killTrackerSocketTypeHash,
@@ -7,6 +6,7 @@ import {
 import { RootState, ThunkDispatchProp } from 'app/store/types';
 import { DestinySocketCategoryStyle } from 'bungie-api-ts/destiny2';
 import clsx from 'clsx';
+import { SocketCategoryHashes } from 'data/d2/generated-enums';
 import React, { useState } from 'react';
 import ReactDOM from 'react-dom';
 import { connect } from 'react-redux';
@@ -16,7 +16,7 @@ import { DimItem, DimPlug, DimSocket } from '../inventory/item-types';
 import { inventoryWishListsSelector } from '../wishlists/selectors';
 import { InventoryWishListRoll } from '../wishlists/wishlists';
 import './ItemSockets.scss';
-import Plug from './Plug';
+import Socket from './Socket';
 import SocketDetails from './SocketDetails';
 
 interface ProvidedProps {
@@ -25,9 +25,6 @@ interface ProvidedProps {
   minimal?: boolean;
   updateSocketComparePlug?(value: { item: DimItem; socket: DimSocket; plug: DimPlug }): void;
   adjustedItemPlugs?: DimAdjustedItemPlug;
-  /** Extra CSS classes to apply to perks based on their hash */
-  classesByHash?: { [plugHash: number]: string };
-  onShiftClick?(lockedItem: LockedItemType): void;
 }
 
 interface StoreProps {
@@ -51,9 +48,7 @@ function ItemSocketsGeneral({
   item,
   minimal,
   wishlistRoll,
-  classesByHash,
   isPhonePortrait,
-  onShiftClick,
   updateSocketComparePlug,
   adjustedItemPlugs,
 }: Props) {
@@ -124,9 +119,7 @@ function ItemSocketsGeneral({
                     isPhonePortrait={isPhonePortrait}
                     socket={socketInfo}
                     wishlistRoll={wishlistRoll}
-                    classesByHash={classesByHash}
                     onClick={handleSocketClick}
-                    onShiftClick={onShiftClick}
                     adjustedPlug={adjustedItemPlugs?.[socketInfo.socketIndex]}
                   />
                 )
@@ -168,56 +161,4 @@ function categoryStyle(categoryStyle: DestinySocketCategoryStyle) {
     default:
       return null;
   }
-}
-
-function Socket({
-  defs,
-  item,
-  socket,
-  wishlistRoll,
-  classesByHash,
-  isPhonePortrait,
-  onClick,
-  onShiftClick,
-  adjustedPlug,
-}: {
-  defs: D2ManifestDefinitions;
-  item: DimItem;
-  socket: DimSocket;
-  wishlistRoll?: InventoryWishListRoll;
-  /** Extra CSS classes to apply to perks based on their hash */
-  classesByHash?: { [plugHash: number]: string };
-  isPhonePortrait: boolean;
-  onClick(item: DimItem, socket: DimSocket, plug: DimPlug, hasMenu: boolean): void;
-  onShiftClick?(lockedItem: LockedItemType): void;
-  adjustedPlug?: DimPlug;
-}) {
-  const hasMenu = Boolean(!socket.isPerk && socket.socketDefinition.plugSources);
-
-  return (
-    <div
-      className={clsx('item-socket', {
-        hasMenu,
-      })}
-    >
-      {socket.plugOptions.map((plug) => (
-        <Plug
-          key={plug.plugDef.hash}
-          plug={plug}
-          item={item}
-          socketInfo={socket}
-          defs={defs}
-          wishlistRoll={wishlistRoll}
-          hasMenu={hasMenu}
-          isPhonePortrait={isPhonePortrait}
-          className={classesByHash?.[plug.plugDef.hash]}
-          onClick={() => {
-            onClick(item, socket, plug, hasMenu);
-          }}
-          onShiftClick={onShiftClick}
-          adjustedPlug={adjustedPlug}
-        />
-      ))}
-    </div>
-  );
 }
